Exception display method can be customized.
